refactor(component-library): align useMeteorI18n file naming with the other composables
