+++ packages/component-library/src/components/_internal/mt-label.vue
@@ -36,7 +36,7 @@ const messages = {
 <script setup lang="ts">
 import { computed, useAttrs } from "vue";
 import MtIcon from "../mt-icon/mt-icon.vue";
-import { useMeteorI18n } from "@/composables/use-meteor-i18n";
+import { useMeteorI18n } from "@/composables/useMeteorI18n";
 
 const props = withDefaults(
   defineProps<{
