+++ packages/component-library/src/components/mt-banner/mt-banner.vue
@@ -51,7 +51,7 @@ import MtIcon from "../mt-icon/mt-icon.vue";
 import MtText from "@/components/mt-text/mt-text.vue";
 import MtButton from "@/components/mt-button/mt-button.vue";
 import { useFutureFlags } from "@/composables/useFutureFlags";
-import { useMeteorI18n } from "@/composables/use-meteor-i18n";
+import { useMeteorI18n } from "@/composables/useMeteorI18n";
 
 const { t } = useMeteorI18n({
   namespace: "mt.banner",
