+++ packages/component-library/src/components/mt-data-table/sub-components/mt-data-table-reset-filter-button/mt-data-table-reset-filter-button.vue
@@ -16,7 +16,7 @@ const messages = {
 </script>
 
 <script setup lang="ts">
-import { useMeteorI18n } from "@/composables/use-meteor-i18n";
+import { useMeteorI18n } from "@/composables/useMeteorI18n";
 
 defineProps<{
   numberOfAppliedFilters: number;
