+++ packages/component-library/src/components/mt-promo-badge/mt-promo-badge.vue
@@ -23,7 +23,7 @@ const messages = {
 
 <script setup lang="ts">
 import { computed } from "vue";
-import { useMeteorI18n } from "@/composables/use-meteor-i18n";
+import { useMeteorI18n } from "@/composables/useMeteorI18n";
 import MtBadge from "../mt-badge/mt-badge.vue";
 
 const props = withDefaults(
