+++ packages/component-library/src/components/mt-text-editor/_internal/mt-text-editor-toolbar-button-table.vue
@@ -57,7 +57,7 @@ import mtSwitch from "@/components/mt-switch/mt-switch.vue";
 import mtTextEditorToolbarButton from "./mt-text-editor-toolbar-button.vue";
 import mtNumberField from "@/components/mt-number-field/mt-number-field.vue";
 import mtModalClose from "@/components/mt-modal/sub-components/mt-modal-close.vue";
-import { useMeteorI18n } from "@/composables/use-meteor-i18n";
+import { useMeteorI18n } from "@/composables/useMeteorI18n";
 import { meteorTextEditorMessages } from "../mt-text-editor.i18n";
 
 const { t } = useMeteorI18n({ messages: meteorTextEditorMessages });
